Comment in of model blocks with an explicit writing of all the equations
The equations which were written through the use of lists |O| for Domestic (D) and Importations (M), as |V| (where V stands for a model variable) are replaced by the explicit equations.

The objective is to ease the tractability of the code by fully expressing each model variable by its name.