product level VAT rates added
Instead of using a flat VAT rate from OECD, we now use national product level VAT rates (European Commission, VAT rates applied in the Member States of the European Union, Taxud.c.1(2018) – EN). The national VAT rates are scaled to match the OECD rates. Note that the OECD rate of 13.74% is not entirely respected because the rates are capped at 20%. The resulting effective VAT rate is therefore 13.65% instead.